New: Building the control unit infrastructure
Put in place a first version of the control unit reading the instruction
RAM and feeding the ALU. An unit test suite is under develpment to
ensure its correctness.

The control unit for the moment contains the instruction decoder, a FIFO
to store the ALU instruction and a state machine to process the incoming
instruction.