Change: Replace rand_sequence and latency testbenchs by a friscv_rv32i testbench
After finding a bug in the control unit, it appears latency and
rand_sequence weren't able to find it. The development of the RV32i
testbench will help to better test the control unit and more, while
aiming to stress the module at system level, emulating the integration
in a complete processor system. To develop once the ALU will be almost
alive.